+++ src/c/v4.c
@@ -8,7 +8,7 @@
 #define SUM_MODE 1
 
 extern void print_csr(uint32_t *, uint32_t *, uint32_t, uint32_t);
-void spmv(uint64_t*, uint32_t*, uint32_t*, uint32_t*, const uint32_t, const uint32_t);
+void spmv(uint32_t*, uint32_t*, uint32_t*, uint32_t*, const uint32_t, const uint32_t);
 int binary_search (uint32_t*, uint32_t, int32_t, int32_t);
 uint32_t sum_common(uint32_t, uint32_t, uint32_t*, uint32_t*);
 
@@ -19,7 +19,7 @@ uint32_t sum_common(uint32_t, uint32_t, uint32_t*, uint32_t*);
  * The lower triangular will be used to walk only the half matrix
  *
  */
-void v4(uint64_t *vertices, uint32_t *csc_row_complete, uint32_t *csc_col_complete, uint32_t *csc_row_low, uint32_t *csc_col_low,
+void v4(uint32_t *vertices, uint32_t *csc_row_complete, uint32_t *csc_col_complete, uint32_t *csc_row_low, uint32_t *csc_col_low,
             const uint32_t nnz_complete, const uint32_t n) {
     printf("\n----------Version 4 is called----------\n");
 
@@ -124,7 +124,7 @@ int binary_search(uint32_t *array, uint32_t key, int32_t low, int32_t high) {
  * We divide the values by two to find the correct number of triangles.
  *
  */
-void spmv(uint64_t *y, uint32_t *csc_row, uint32_t *csc_col, uint32_t *values, const uint32_t nnz, const uint32_t n) {
+void spmv(uint32_t *y, uint32_t *csc_row, uint32_t *csc_col, uint32_t *values, const uint32_t nnz, const uint32_t n) {
 
     // x vector will be always 1, so change x -> 1
     for(uint32_t i = 0; i<n; i++) {
